restructure oAuth clients folder: example for other folders
