+++ apps/api/v2/src/modules/auth/guard/oauth/access-token.guard.ts
@@ -1,4 +1,4 @@
-import { OAuthFlowService } from "@/modules/oauth/flow/oauth-flow.service";
+import { OAuthFlowService } from "@/modules/oauth-clients/services/oauth-flow.service";
 import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from "@nestjs/common";
 
 @Injectable()
