+++ apps/api/v2/src/modules/endpoints-module.ts
@@ -1,11 +1,10 @@
 import { BookingModule } from "@/modules/booking/booking.module";
-import { OAuthFlowModule } from "@/modules/oauth/flow/oauth-flow.module";
-import { OAuthClientModule } from "@/modules/oauth/oauth-client.module";
+import { OAuthClientModule } from "@/modules/oauth-clients/oauth-client.module";
 import type { MiddlewareConsumer, NestModule } from "@nestjs/common";
 import { Module } from "@nestjs/common";
 
 @Module({
-  imports: [BookingModule, OAuthClientModule, OAuthFlowModule],
+  imports: [BookingModule, OAuthClientModule],
 })
 export class EndpointsModule implements NestModule {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
