+++ apps/api/v2/src/modules/oauth-clients/inputs/update-oauth-client.ts
