+++ apps/api/v2/src/modules/oauth-clients/oauth-client.module.ts
@@ -1,0 +1,29 @@
+import { getEnv } from "@/env";
+import { AuthModule } from "@/modules/auth/auth.module";
+import { MembershipModule } from "@/modules/membership/membership.module";
+import { OAuthClientUsersController } from "@/modules/oauth-clients/controllers/oauth-client-users/oauth-client-users.controller";
+import { OAuthClientsController } from "@/modules/oauth-clients/controllers/oauth-clients/oauth-clients.controller";
+import { OAuthFlowController } from "@/modules/oauth-clients/controllers/oauth-flow/oauth-flow.controller";
+import { OAuthClientGuard } from "@/modules/oauth-clients/guards/oauth-client/oauth-client.guard";
+import { OAuthClientRepository } from "@/modules/oauth-clients/oauth-client.repository";
+import { OAuthFlowService } from "@/modules/oauth-clients/services/oauth-flow.service";
+import { PrismaModule } from "@/modules/prisma/prisma.module";
+import { TokensRepository } from "@/modules/tokens/tokens.repository";
+import { UserModule } from "@/modules/user/user.module";
+import { Global, Module } from "@nestjs/common";
+import { JwtModule } from "@nestjs/jwt";
+
+@Global()
+@Module({
+  imports: [
+    PrismaModule,
+    AuthModule,
+    UserModule,
+    MembershipModule,
+    JwtModule.register({ secret: getEnv("JWT_SECRET") }),
+  ],
+  providers: [OAuthClientRepository, OAuthClientGuard, TokensRepository, OAuthFlowService],
+  controllers: [OAuthClientsController, OAuthClientUsersController, OAuthFlowController],
+  exports: [OAuthClientRepository, OAuthClientGuard],
+})
+export class OAuthClientModule {}
